refactor: extracted computer icon :recycle:
