+++ src/App.tsx
@@ -45,17 +45,11 @@ import PATHS from "commons/constants/PATHS";
 import NotFoundPage from "./features/404/NotFound.page";
 import useCheckApiConnection from "./commons/hooks/connection/useCheckApiConnection";
 import scanMenuPage from "./features/computer/ScanMenu/ScanMenu.page";
+import {ComputerIconComponent} from "./commons/components/Icons";
 
 setupIonicReact();
 
-function ComputerIconComponent() {
-    return <svg width="40" height="41" viewBox="0 0 40 41" fill="none" xmlns="http://www.w3.org/2000/svg">
-        <rect x="0.5" y="1" width="39" height="39" rx="19.5" fill="#193549" stroke="#CCCCCC"/>
-        <path
-            d="M26.6667 14.5C27.1087 14.5 27.5326 14.658 27.8452 14.9393C28.1577 15.2206 28.3333 15.6022 28.3333 16V23.5C28.3333 24.3325 27.5917 25 26.6667 25H30V26.5H10V25H13.3333C12.8913 25 12.4674 24.842 12.1548 24.5607C11.8423 24.2794 11.6667 23.8978 11.6667 23.5V16C11.6667 15.1675 12.4083 14.5 13.3333 14.5H26.6667ZM26.6667 16H13.3333V23.5H26.6667V16ZM20 20.5C21.8417 20.5 23.3333 21.175 23.3333 22V22.75H16.6667V22C16.6667 21.175 18.1583 20.5 20 20.5ZM20 16.75C20.442 16.75 20.8659 16.908 21.1785 17.1893C21.4911 17.4706 21.6667 17.8522 21.6667 18.25C21.6667 18.6478 21.4911 19.0294 21.1785 19.3107C20.8659 19.592 20.442 19.75 20 19.75C19.075 19.75 18.3333 19.0825 18.3333 18.25C18.3333 17.4175 19.0833 16.75 20 16.75Z"
-            fill="#CCCCCC"/>
-    </svg>;
-}
+
 
 function CodebarIconComponent() {
     return <svg width="40" height="41" viewBox="0 0 40 41" fill="none"
