+++ src/commons/components/Icons/index.tsx
@@ -1,3 +1,4 @@
 import AlertIconComponent from "./AlertIcon/AlertIcon.component";
+import ComputerIconComponent from "./ComputerIcon/ComputerIcon.component";
 
-export {AlertIconComponent};
+export {AlertIconComponent, ComputerIconComponent};
